+++ Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/AccelByteUe4Sdk.Build.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2019 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2018-2020 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
@@ -26,7 +26,8 @@ public class AccelByteUe4Sdk : ModuleRules
             "WebSockets",
             "Networking",
             "SSL",
-            "Icmp"
+            "Icmp",
+            "Agones"
         });
 
         if (Target.bBuildEditor == true)