+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.17.0 (2020-06-19)
+### Added
+- Agones SDK UE4 plugin
+- Register server using Agones provider
+
 ## 2.16.0 (2020-06-16)
 ### Added
 - Agreement Api